✨ feat: add fonts
- add fonts
- timer clock lcd output dart file
- change new font timer clock
